requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.3
distlib==0.3.1
Django==3.0.8
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-filter==2.3.0
djangorestframework==3.11.1
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
pep8==1.7.1
Pillow==7.2.0
protobuf==3.12.2
pycodestyle==2.6.0
pylint==2.5.3
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.27
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
